RESOLVED - bug 185162: [C99 parser] no support for hex and character constants in preprocessor conditional expressions https://bugs.eclipse.org/bugs/show_bug.cgi?id=185162
Patch from Mike Kucera

+++ c99/org.eclipse.cdt.core.parser.c99/src/org/eclipse/cdt/internal/core/dom/parser/c99/C99ExprEvaluator.java
@@ -183,7 +183,7 @@ public void addToken(IToken token) {
     token.setAdjunctIndex(adjuncts.size());
 }
 
-public Integer evaluate() {
+public Long evaluate() {
 	parser(null, -1);
 	return action.result();	
 }
